refactor: move functions into helpers
